Implement unified search_resources method
We have a lot of search_XXX calls in the cloud layer, but not for everything.
They are also doing nearly same to what it is possible to do with plain proxy 
methods and therfore we want to simplify and unify this so that Ansible modules
can easily rely on a single function in different modules instead or needing a 
dedicated search method for every resource doing the same.

New method accepts resource_type, resource identifier (which may be
empty), filters and also possibility to pass additional args into the
_get and _list calls (for unpredicted special cases). With this all
search_XXX functions can be finally deprecated.

Change-Id: I375c2b625698c4920211eb6e089a1b820755be84