compute: don't pass networks: auto for older microversions
Change I1a6ba311ddedc1b8910051257299d3acd367df46 made this
unconditionally specify "networks: auto" when no network info was
specified.  From the nova release notes

 "Starting with microversion 2.37, this field is required and the
 special string values auto and none can be specified for
 networks. auto tells the Compute service to use a network that is
 available to the project, if one exists."

So we should restrict where this is specified.

This was noticed when trying to use openstacksdk >=0.99 against the
Rackspace API.  I have run this change against that API and verified
it is not passing the argument.

Change-Id: Ifc5bd0ddc92bbbc7b250df3f3931518f62cc339f