Revert "Revert "Add "security_group_ids" to Port's query parameters""
This reverts commit 5e969fb49d96fd2071ecc8a379ec6e2c3c5c90bc.

Reason for revert: The revert happened to resolve a merge conflict and was supposed to be re-reverted afterwards. This is the re-revert.

Original Commit message for original change Ic557dc3bf97a193fd28c13792302fb8396e29df1:

Neutron's port-security-groups-filtering extension allows to filter
ports by security-group(s). This feature was added with [1].

We keep the name like in the attribute "security_group_ids" instead of
"security_groups" to stay consistent within the Port model.

[1] https://launchpad.net/neutron/+bug/1405057

Change-Id: I060edb9c0d2eeb03d514707b69978f3d5f0ae3f4