image: Remove unsupported parameters from v1 proxy
There was an idea to provide the exact same method signature for
'create_image' in the v1 and v2 Image proxy APIs. This was a bad idea as
it limits how the v2 version of this method can evolve. Stop doing it.
Or rather, try. In reality, nothing really changes and all we're doing
is removing the documentation for these unsupported options and
reworking the exception logic.

The docstring for the 'create_image' function is reworked.

Change-Id: I65da2e6795d7fd417dd2ab46c272b36e4ae83b69
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>