Update tox.ini for tox v4 compatibility
Tox v4 made a number of changes that break compatibility with old tox v3
tox.ini files. In particular openstacksdk lists env vars for passenv on
a single line separating them with spaces. This is no longer allowed and
produces this error:

  failed with pass_env values cannot contain whitespace, use comma to have multiple values in a single line, invalid values found 'OS_* OPENSTACKSDK_*'

We can fix this by listing each value on a single line which we do here.

Depends-On: https://review.opendev.org/c/openstack/python-openstackclient/+/868430
Change-Id: Ibf4a16e544921ec914f6e23112e0fc37a12096a5