config: Load additional options for v3multifactor
As the name would suggest, 'v3multifactor' uses multiple factors for
authentication. As a result, we need to register the configuration
options for each required auth method. If we don't do this, we won't
move the required configuration options for auth out of the top level
config object into the 'auth' key. This affects users who are
configuring via environment variables or config options.

Normally registering of configuration options is handled by the
'MultiFactor.load_from_options' [1][2] method but there doesn't appear
to be a way to "register" the auth methods without actually loading the
plugin. As a result, if we encounter this auth type then we need to do
this registration of extra options manually. There's a probably a TODO
for keystoneauth to provide a mechanism for this but I don't know what
that would look like right now.

[1] https://github.com/openstack/keystoneauth/blob/5.1.2/keystoneauth1/loading/_plugins/identity/v3.py#L332-L340
[2] https://github.com/openstack/keystoneauth/blob/5.1.2/keystoneauth1/loading/_plugins/identity/v3.py#L323-L329

Change-Id: I1f02133be373fa1f8facfd016586395fa2379a3e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Story: 2010661
Task: 47714