image: Don't envelope properties
The Glance v2 API is weird, in that it allows and places arbitrary
properties in the root envelope of requests and responses respectively
rather than under a 'properties' or 'metadata' key like other services.
We were using a properties key temporarily as an envelope so we can
format these extra properties specifically. However, we were forgetting
to remove this envelope when creating an bare image without uploading
data. Correct this.

Change-Id: Idd2dd85e2fee364348328b5f25096f565cd47241
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Story: 2009127
Task: 43041