Use pre-commit for 'pep8' tox target, bump versions
We want to bump the versions of hacking, but doing so requires changes
in two places: '.pre-commit-config.yaml' and 'tox.ini'. This is silly:
we can simply use tox to handle pre-commit and leave all other
dependencies to pre-commit. Do this, migrating doc8 to pre-commit and
bumping the other dependencies in the process.

Change-Id: I26fa07145129d3ef9cb17693427ed70e55dbaaf5
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>