Add fake resources generator
In a variety of places we need to have an easy way to have resources
populated with fake data. It costs quite a lot of lines of code and can
be easily automated. With this change it can be as easy as `_fake =
sdk_fakes.generate_fake_resource(Project)`. The code is implemented not to
require estabilshed connection to ease use in tests.

Change-Id: I47312f4036a0b389cd3689466ab220ba558aa39a