Allow tags to be passed through to compute.create_server
Setting tags on servers is useful for managing sets of related resources
(e.g., we're spawning JupyterHub instances in OpenStack and we want to
tag all servers created by our spawner to make it easier to clean things
up).

This commit configures the create_server method to accept the tags
keyword paramter and pass it through to compute.create_server.

Change-Id: I92a449e584f9912f2bff86939e3565c6b3001bb9