cloud: Ignore invalid filters
The 'openstack.cloud._utils._filter_list' function accepts a list of
filters. These filters are key-value pairs where the key indicates the
attribute of the resources that we wish to filter on, and the value is
either a simple value or another key-value pair to allow us to filter on
nested attributes (turtles all the way down 🐢🌍).

We were using '.get(key, None)' to retrieve the attribute from the
resource, meaning we inadvertently allowed for filters on attributes
that did not exist. If the attribute was not found, we treated this as a
mismatch. This is certainly not what was intended and it's not
necessary: because the cloud layer is now using the proxy layer rather
than raw REST requests, and the proxy layer in-turn uses resources, we
can ensure that a valid attribute will always be present even if it
unset. As such, the current logic is incorrect and has allowed at least
one bug to slip into our codebase.

Correct this issue by checking for the presence of a resource attribute
matching the key. If an attribute doesn't match, we now raise a warning
and exception, as one would have expected.

Change-Id: Id8756f679ec5130779d61d454455c9c25afe8baa
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Story: 2010627
Task: 48095