cloud: Remove '_object_store_client'
Replace the use of a raw client with a REST request using our proxy
clients. These things aren't identical since the use of the proxy
clients necessitates a whole load of version negotiation but it's close
enough and let's us dump the whole raw client idea.

Note that we need to do a little manipulation of some tests. We had
tests that were manipulating 'time.time()'. This was conflicting with
the rate limiting semaphore provided in keystoneauth. By mocking a
wrapper function instead, we can make our mock a little more targeted.

Change-Id: I39d171e4456ea491577cb46216c321da2c9d2d19
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>