fix connection.Connection finalizer
finalizers execution order is not guaranteed in python3. This leads to
absence of builtins like "open" in the finalizers. Depending on the
keyring backend used inside of the set_auth_cache access to the file system
may be required.

Replace use of destructor to another builtin library "atexit" to
implement a more reliable exit handler.

Change-Id: I2d6882d64b57b65dbef086a08a3b3dc6b5faa2be