Replace appdirs usage with platformdirs
The appdirs library has been abandoned for several years now, and a
fork named platformdirs is the recommended alternative. It's almost
a drop-in replacement, except the module and convenience class name
have changed. Adjust the internal variable name we're using for the
expansion to match as well, for all those consistent hobgoblins.

Add it with a >=3 lower bound because that introduces a breaking
change of the default configuration location for macOS users.
Include a release note about that too.

Depends-On: https://review.opendev.org/886642
Change-Id: I4c2b60ca0da29da1f605b4c0b5e77f8e1071a19c