Return HTTP response for delete_access_rule.
In the original code, when ignore_missing parameter is set to False,
no information is reported when either the share or the access is
not found. However, after applying this patch, the HTTP response
is returned to users, allowing them to verify whether the request
completed successfully or not. If there are serious issues with
the request clients, an exception will be raised to handle those
situations appropriately.

Change-Id: I925d943a00b4de198013336f5043753552255247