mypy: Address issues with openstack.block_storage
Our first significant bug! We were registering the 'group_type' module
rather than the 'GroupType' resource in our block storage v3 proxy's
resource registry. This is corrected. The rest of the changes are simply
to work around mypy (obviously) not being able to detect the key and
value types for empty dicts. Cinder doesn't care about the values of the
dict for most WSGI actions so this doesn't matter.

Change-Id: I4d9a4d4fcf08e8aa13b47314b24b2bb7226be3eb
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>