fix memory leak of Connections
during `__init__` of a `Connection` we register an `atexit` handler to
call the `close` method when the python interpreter exits. This should
ensure that connections are closed reliably.

However when a connection is closed earlier (e.g. by using the context
manager of `Connection`) `close` is already called. However the `atexit`
handler is not unregistered.
This prevents python garbage collector from actually collecting the
`Connection` since there is still an active reference to it.

To fix this we unregister the `atexit` handler if `close` is called.

Change-Id: Id02a495353e1bfef3785248e48dec7344ef8f5af