Fix: Update/Delete type encryption operation
When requesting to update/delete the volume type encryption,
we do a GET encryption type call and fail with KeyError: 'volume_type_id'.
This is because the request URL for GET encryption type is created as

'/types/%(volume_type_id)s/encryption'

and we are passing 'volume_type' instead of 'volume_type_id'.
This patch fixes the issue by passing 'volume_type_id'.

Story: #2010896
Task: #48686
Change-Id: Ic3d528e0e7b8e07a72f83cbe84442dc17ae32c27