cloud: Remove remnants of server caching
The final cleanup post I2c8ae2c59d15c750ea8ebd3031ffdd2ced2421ed. Once
again, this is all dead code now that can be removed in favour of proxy
layer caching.

Note that the removed test can be removed since this is a test for this
cloud-layer caching, not the proxy-layer caching (which would require
configuration of dogpile.cache).

Change-Id: Ifb80965a8c4faa53cbdbcb1e1efd8a5b90ddb0e7
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>