[baremetal] Add support for parent_nodes
Ironic introduced the concept of parent_node in the bobcat
development cycle where a nested relationship exists for
the purposes of management of nodes and related interactions.

This change adds the field on the SDK and allows the field
to be queried utilizing the additional features added to the
API for listing child nodes and matching nodes with a parent
node.

Change-Id: I415cf2049e31821a183eed3b17e3701eaa5dd23f