Enforce endpoint override for services without discovery
Services without version discovery implemented are tricky. When user is
not using endpoint_override for those there are sporadic and often
attempts to perform version discovery on every request sent [1]. It
could be prevented as much as possible on SDK side during proxy
initialization, but when a request is sent to the proxy this is still
happening inside the keystonauth library. Forcing endpoint_override for
those help avoiding that.

Touching cloud_region triggers mypy complain, so change urllib import to
keep it happy.

[1]
https://lists.openstack.org/archives/list/openstack-discuss@lists.openstack.org/thread/GBBHV2LN5S3HEUQHUQGZZ3UFEVRHQBOS/

Change-Id: I1956dfaac9ba8e3f9234cc98a36b20a330dc88f6