Fix AttributeError in delete_limit method
This commit resolves an issue in the delete_limit method of the
Identity V3 Proxy class. The method was incorrectly trying to
access a 'Limit' attribute of a 'Limit' object, causing an
AttributeError. The fix involves correctly passing the 'Limit'
class from the '_limit' module as the first argument to the
'_delete' method.

Closes-Bug: #2052650
Change-Id: I435f52f8085c899fe1be57fcd5a2ce6f04618427