tox: Correct functional test factors
We are running functional tests in zuul without a 'pyNN' factor (e.g.
'tox -e functional'). For this to work, we need to allow an empty
factor, i.e. we want:

  [testenv:functional{,-py310}]

rather than:

  [testenv:functional{-py310}]

(note the missing comma)

Unfortunately we missed this as tox 4 has a currently unaddressed
regression [1] that results in it running the base testenv in the case
there is only a partial factor match. That needs to be fixed for avoid
this biting us again the future, but we can at least fix it for now.

[1] https://github.com/tox-dev/tox/issues/3219

Change-Id: Ib9f65a4523222f1224d51534c5061f90501b59d3
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>