Allow filtering BGPVPNs
BGPVPNs can be filtered server-side when being listed. The most
interesting filters here are for routers, networks and ports, as these
allow to filter for a specific association the BGPVPN has, but the other
filter options might come in handy as well.

route|import|export-target support only exact matches in the BGPVPN API,
meaning all entries need to be given in the right order as a comma
separated string. As this is quiet counter-intuitive for a user (that
might expect they could provide only one or unsorted entries) these
attributes have been left out.

Change-Id: I355b6584e115cf70a7fd040a24d30f9686555b85