cloud: Combine networking-related classes
The SecurityGroupCloudMixin, FloatingIPCloudMixin and
NetworkingCommonCloudMixin classes all exist to handle the difference
between nova-network and neutron. Combine them and use them as a
subclass for the ComputeCloudMixin and NetworkingCloudMixin classes.
This makes our class hierarchy a little messier but it's the only
obvious way I have to address this issue that doesn't involve something
like decorator gymnastics.

Change-Id: I4b101124e19636dfa9a0f704914f5100870a9fb9
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>