zuul: Use more stable dib job
centos-8-stream is EOL and the builds are no longer passing. Since the
only things that matters here is how the sdk is being used by nodepool,
the actual image being built doesn't really matter. Use Ubuntu Noble.
which is the most recent addition and thus will still be working for a
long time, or so we hope.

Also stop running non-voting jobs in the gate pipeline.

Change-Id: I351ce792ffd1d24d032d7a657366be807ea29e38