Add timeline for removal of deprecated features
Not all deprecated features are deprecated for removal, but those that
are need a hint for when they will start being removed. Start making use
of the new class of version-specific deprecation warnings added in
change I8bb0bfb039593c5b59f1f9c16523a090d899f099 to set timers for the
various deprecated-for-removal features we have.

Change-Id: I17d6703e82d49fca9743571f2bf1ba9c1fb3d190
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>