resource: Rework fields
The Body, Header, URI and Computed field types are not values in their
own right: rather, they are indicators of where the real values will
come from. At some point, we should probably make better use of type
annotations but for now, modify things so that 'openstack.resource.Body'
et al are functions that return 'Any' and don't conflict with explicit
type annotations. This allows us to change some of the attributes of the
base resource - namely the 'name' and 'location' fields - to reflect
their "real" types.

Change-Id: Ia771d18cf6d8d72dbf995418166e996515f5d52c
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>