placement: Add support for resource provider allocations
Once again, placement is being different. Unlike most OpenStack APIs
that returns a list of objects, the resource provider allocation list
operation returns an object with consumer IDs as the keys and
allocations (nested under a 'resources' key) as the values. Put another
way, we see:

  {
    "56785a3f-6f1c-4fec-af0b-0faf075b1fcb": {
      "resources": {
        "MEMORY_MB": 256,
        "VCPU": 1
      }
    }
  }

instead of the more typical:

  [
    {
      "consumer_id": "56785a3f-6f1c-4fec-af0b-0faf075b1fcb",
      "resource_class": "MEMORY_MB",
      "usage": 256
    },
    ...
  ]

This need special handling code, namely a reimplementation of the 'list'
class method. Thankfully, we can only list allocations so no other
methods are needed.

Change-Id: I6e36cf7b7630ff57bd4bb314f502d125f2f409ff
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>