proxy: Remove '_check_resource' decorator
This was never being configured with strict=true, meaning the logic
can be moved into the '_get_resource' helper.

Fixing this highlights a few places where we need to be a little more
permissive in our proxy function signatures, so these are adjusted along
with the doc strings for same.

Change-Id: Iba51d19ddb1420f297bc79402f008512c041a721
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>