Deprecate 'use_direct_get' parameter
This is additional complexity that we don't want to support and won't be
able to support once everything in the cloud layer is switched over to
use the proxy layer. If users want control over this, they can use the
'get_xxx' methods in the proxy layer instead of the cloud layer.

Change-Id: I76e9e2873ad57db673516c2b66f9280f6b0280be
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>