Fix volume summary passing non-existent all_projects
The parameter all_projects is invalid and the
correct parameter to use is all_tenants.

We also fix so that more parameter can be
passed down to the GET request for filtering.

This was found when fixing so that this API
call actually passes the parameters down. [1]

[1] https://review.opendev.org/c/openstack/cinder/+/928781/1

Change-Id: Iea983a60212d1e950ce78b1d89edcaa822d9458e
(cherry picked from commit b7e8e869fa6641be3e4d41a3b53e93e0c9d005d8)