fields: Tweak _convert_type
We would like to do a larger rework of this, but for now simply reflow
information somewhat to make that larger rework a little more
understandable.

The main functional change here is the removal of an 'isinstance(value,
data_type)' branch that would skip casting of the value if the value was
already of the expected type. This is removed since it provides limited
value (e.g. int(5) == 5, str('test') == 'test') and would make the
future rework slightly more complex.

We also add lots of tests to prevent regressions in future changes.

Change-Id: Icb6b26127d8e200b650687578f450490e71d24c3
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>