typing: Annotate openstack.fields (1/2)
This focuses solely on annotating the '_convert_value' helper, on
account of the size of the annotations. This size is a side effect of
the amount of potential data_type types we can receive, each of which
will result in a slightly different return value. These are all called
out and hopefully make some sense when taken individually.

Change-Id: I812d36142e90a1f0f38619ece76ce772f671acc7
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>