Fix TestKeypairAdmin tests and compute.*_keypair
The get_keypair and delete_keypair calls need to specify the user_id if
the keypair belongs to another user. This was only working by accident
because the first element returned by list_users was the current user.
However, the ordering of the user list is now different following a
change in Keystone [1].

This failing test uncovered additionally that keypair methods are not
properly setting user_id into the query parameters. Overload fetch and
delete methods to properly pass user_id as a parameter.

[1] https://review.opendev.org/c/openstack/keystone/+/938814

Closes-Bug: #2095312
Co-Authored-By: Artem Goncharov <artem.goncharov@gmail.com>

Change-Id: Ic552dee83d56278d2b866de0cb365a0c394fe26a