cloud: Replace use of cloud methods in identity cloud layer
Always use the proxy layer methods instead.

We also change the logic of the '_get_grant_revoke_params' helper. This
previously returned the same error about a user needing to be specified
if (a) a user was not specified or (b) a user was specified but it was
invalid. (b) makes no sense and as such no longer happens.

Change-Id: I81a6ed6143c43eea1ba7202db6d54f3859852a53
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>