typing: Add (more) missing proxy arguments
The Proxy._update method also requires a value argument. Once again,
this currently works only because the '_check_resource' decorator is
filling in the missing value for us.

mypy didn't/couldn't pick this up because of our use of untyped
'**attrs' in both the base class and all callers. Another one we'd like
to resolve long-term but haven't managed to do yet.

Change-Id: I73920092a474744f1d7b154d36bc2af1aeb5eb53
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>