[stable-only] Fix gate (cap os-client-config and fix doc job)
This is a combination of 2 commits:

1. Cap os-client-config

os-client-config became a cycle-independent project without stable and
unmaintained branches, hence it is taken from master branch always.
This broke the gate as os-client-config recently dropped support for
python version < 3.10.

This patch fixes the gate by pinning os-client-config to the previous
release that still has the support.

2. docs: Correct docs for VpnIpsecPolicy

We were incorrectly documenting the VpnIkePolicy twice instead.

While we're here, add a docstring for the VpnIkePolicy resource so the
documentation is actually useful.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
(cherry picked from commit 3bbbf1abdf9fe769f31f5934cea95514619c03e4)

Change-Id: I032fe8c28a88a31039770fc031ee7d8191269561
Signed-off-by: Elod Illes <elod.illes@est.tech>