tests: Use binary mode to open files
A recent change to the CONTRIBUTING.rst file introduced unicode
characters, which requests does not handle correctly when the file is
open in text mode. requests 3.0 is removing support for opening files in
text mode so we kill two birds with one stone here by switching to
opening the file in binary mode. We also remove the unicode character (I
tested the fix separately), correct the project name in the document,
and use the raise from syntax to get better error handling in these
cases.

Change-Id: I47d65383233e23ec5edf1e6735c39c66eeafb804
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>