ci: ironic: don't use tinyipa
Ironic CI jobs, espescially when we wanted to reduce the memory
footprint being used, leveraged TinyIPA. Unfortunately, TinyIPA
was always intended with testing and ultimately became harder
and harder to patch because it was a separately maintained and
mirrored project which resulted in a high number of job failures
as well.

This became problematic enough for the Ironic project that it
chose to intentionally deprecate and remove tinyipa.

And surprise, the job is still here trying to use it.
Which means this change also needs to be backported to 2025.2
at a minimum.

Change-Id: I5d3699312e4f2243b8c43e0e9f8c5624b433ad89
Signed-off-by: Julia Kreger <juliaashleykreger@gmail.com>